Add validate_token decorator
