+++ src/views/user.py
@@ -1,34 +1,12 @@
-from functools import wraps
 from flask import Blueprint, request
 from flask_cors import CORS
 
 from controllers import user as controller
 from utils.formatters import create_response
+from utils.validators.general import validate_header
 
 user_blueprint = Blueprint('user', __name__, url_prefix='/api')
 CORS(user_blueprint)
-
-
-def validate_header(func):
-    @wraps(func)
-    def decorated_function(*args, **kwargs):
-        type_error = ('Unsupported media type', 415)
-        header_values = ['*/*', 'application/json']
-        header_keys = ['Accept', 'Content-Type']
-        header = request.headers
-        rm = request.method
-        methods = ['POST', 'PUT', 'PATCH']
-        if rm in methods:
-            if not request.data:
-                return create_response(*type_error)
-            if not all(h in header for h in header_keys):
-                return create_response(*type_error)
-            if not all(header[h] in header_values for h in header_keys):
-                return create_response(*type_error)
-
-        return func(*args, **kwargs)
-
-    return decorated_function
 
 
 @user_blueprint.route('/users/', methods=['GET', 'POST'])
@@ -51,7 +29,9 @@ def user_by_username(user_username):
         response, status = controller.get_one_user(user_username)
 
     elif request.method == 'DELETE':
-        response, status = controller.delete_user(user_username)
+        response, status = controller.delete_user(
+                user_username, request.headers
+        )
 
     elif request.method == 'PATCH':
         response, status = controller.update_user(user_username, request.json)
